+++ zarrs/src/array/array_ops/async_array_read_ops_array.rs
@@ -416,6 +416,12 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> AsyncArrayReadOps
         options: &CodecOptions,
     ) -> Result<Option<Vec<u8>>, ArrayError>;
 
+    pub async fn async_retrieve_encoded_subchunk_at_level(
+        &self,
+        level: usize,
+        subchunk_indices: &[u64],
+    ) -> Result<Option<Vec<u8>>, ArrayError>;
+
     pub async fn async_retrieve_encoded_subchunk_at_level_opt(
         &self,
         level: usize,
@@ -423,10 +429,21 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> AsyncArrayReadOps
         options: &CodecOptions,
     ) -> Result<Option<Vec<u8>>, ArrayError>;
 
+    pub async fn async_retrieve_subchunk<T: FromArrayBytes>(
+        &self,
+        subchunk_indices: &[u64],
+    ) -> Result<T, ArrayError>;
+
     pub async fn async_retrieve_subchunk_opt<T: FromArrayBytes>(
         &self,
         subchunk_indices: &[u64],
         options: &CodecOptions,
+    ) -> Result<T, ArrayError>;
+
+    pub async fn async_retrieve_subchunk_at_level<T: FromArrayBytes>(
+        &self,
+        level: usize,
+        subchunk_indices: &[u64],
     ) -> Result<T, ArrayError>;
 
     #[allow(clippy::missing_errors_doc)]
@@ -437,10 +454,21 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> AsyncArrayReadOps
         options: &CodecOptions,
     ) -> Result<T, ArrayError>;
 
+    pub async fn async_retrieve_subchunks<T: FromArrayBytes>(
+        &self,
+        subchunks: &dyn ArraySubsetTraits,
+    ) -> Result<T, ArrayError>;
+
     pub async fn async_retrieve_subchunks_opt<T: FromArrayBytes>(
         &self,
         subchunks: &dyn ArraySubsetTraits,
         options: &CodecOptions,
+    ) -> Result<T, ArrayError>;
+
+    pub async fn async_retrieve_subchunks_at_level<T: FromArrayBytes>(
+        &self,
+        level: usize,
+        subchunks: &dyn ArraySubsetTraits,
     ) -> Result<T, ArrayError>;
 
     #[allow(clippy::missing_errors_doc)]